services/ops-agent/requirements.txt
livekit-agents>=0.12.0
livekit-plugins-aws>=0.2.0
python-dotenv>=1.0.0
